Create android and web specific files and move ios to default
