--- packages/article-skeleton/src/dropcap-util.ios.js
@@ -1,1 +1,0 @@
-export default child => child;
