+++ packages/article-skeleton/src/dropcap-util.web.js
@@ -1,0 +1,3 @@
+import insertDropcapIntoAST from "./dropcap-util-base";
+
+export default insertDropcapIntoAST;
